# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: RailYatra API
  version: 2.0.0
  description: |
    Node.js + SQL Server API for RailYatra train booking and reservation management.
    Master timetable data (Category A) is imported from authorized datasets.
    Fares, availability, RAC/WL (Category B) are development simulations unless noted.
servers:
  - url: /api
paths:
...
  /trains/{id}:
    get:
      summary: Train details (summary, no full route)
      tags: [Trains]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Train summary with classes and running days
        '404':
          description: Not found
...
  /admin/rac/promote:
    post:
      summary: Promote next paid RAC booking to confirmed
      tags: [Admin]
      security: [{ AuthToken: [] }]
      responses:
        '200':
          description: RAC booking promoted
  /admin/data-import/status:
    get:
      summary: Master data import provenance and counts
      tags: [Admin]
      security: [{ AuthToken: [] }]
      responses:
        '200':
          description: Latest imports and master-data counts
  /admin/trains:
    get:
      summary: Paginated admin train list
      tags: [Admin]
      security: [{ AuthToken: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 25 } }
        - { name: search, in: query, schema: { type: string } }
      responses:
        '200':
          description: Paginated trains
  /admin/stations:
    get:
      summary: Paginated admin station list
      tags: [Admin]
      security: [{ AuthToken: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 50 } }
        - { name: search, in: query, schema: { type: string } }
      responses:
        '200':
          description: Paginated stations
components:
  securitySchemes:
    AuthToken:
      type: apiKey
      in: header
      name: x-auth-token